+++ src/cmd/snapshot.rs
@@ -11,12 +11,12 @@ use indexmap::IndexMap;
 use miette::{Context as _, IntoDiagnostic as _};
 use tokio::runtime::Runtime;
 
+use crate::Result;
 use crate::bail;
 use crate::cli::{remove_symlink_dir, symlink_dir};
 use crate::cmd::clone::get_path_subdeps;
 use crate::config::{Dependency, Locked, LockedSource};
 use crate::diagnostic::Warnings;
-use crate::error::*;
 use crate::infoln;
 use crate::sess::{DependencySource, Session, SessionIo};
 use crate::{fmt_path, fmt_pkg, stageln};
