+++ src/cmd/update.rs
@@ -10,11 +10,11 @@ use clap::Args;
 use indexmap::IndexSet;
 use tabwriter::TabWriter;
 
+use crate::Result;
 use crate::cmd;
 use crate::config::{Locked, LockedPackage};
 use crate::diagnostic::Warnings;
 use crate::ensure;
-use crate::error::*;
 use crate::lockfile::*;
 use crate::resolver::DependencyResolver;
 use crate::sess::Session;
