+++ src/lib.rs
@@ -4,7 +4,6 @@ pub mod cli;
 pub mod cmd;
 pub mod config;
 pub mod diagnostic;
-pub mod error;
 pub mod git;
 pub mod lockfile;
 pub mod progress;
@@ -15,3 +14,6 @@ pub mod target;
 pub mod util;
 
 pub use miette::{bail, ensure, miette as err};
+
+pub type Error = miette::Report;
+pub type Result<T> = miette::Result<T>;
