+++ src/lockfile.rs
@@ -6,8 +6,8 @@ use std::path::Path;
 
 use miette::{Context as _, IntoDiagnostic as _};
 
+use crate::Result;
 use crate::config::{Locked, LockedPackage, LockedSource, PrefixPaths};
-use crate::error::*;
 
 /// Read a lock file.
 pub fn read_lockfile(path: &Path, root_dir: &Path) -> Result<Locked> {
