+++ src/sess.rs
@@ -36,12 +36,12 @@ use typed_arena::Arena;
 use crate::cli::read_manifest;
 use crate::config::{self, Config, Manifest, PartialManifest};
 use crate::diagnostic::{Diagnostics, Warnings};
-use crate::error::*;
 use crate::git::Git;
 use crate::progress::{GitProgressOps, ProgressHandler};
 use crate::src::{SourceFile, SourceGroup, SourceType};
 use crate::target::TargetSpec;
 use crate::util::try_modification_time;
+use crate::{Error, Result};
 use crate::{bail, ensure, err};
 
 #[derive(Debug, Error, Diagnostic)]
